refactor ItemCard
doing this to share the component in multiple places -- non-owners
are shown shop controls and the owner is shown item controls with
different delete function depending on the context
